refacto: big refacto for navbar
