+++ frontend/src/legacy/app/app.tsx
@@ -1,7 +1,7 @@
 import { NotificationMenu } from '#modules/notification/view/NotificationMenu/NotificationMenu';
-import { AppMenu } from '#shared/components/PageTemplate/components/AppMenu';
-import { UserMenu } from '#shared/components/PageTemplate/components/UserMenu';
-import { UserMenuUnauthenticated } from '#shared/components/PageTemplate/components/UserMenuUnauthenticated';
+import { AppMenu } from '#shared/components/PageTemplate/AppMenu/AppMenu';
+import { UserMenuAuthenticated } from '#shared/components/PageTemplate/UserMenu/UserMenuAuthenticated';
+import { UserMenuUnauthenticated } from '#shared/components/PageTemplate/UserMenu/UserMenuUnauthenticated';
 
 import { wrapAndRenderLegacyCode } from '../utils/wrapAndRenderLegacyCode';
 import { AppInstallBanner } from './appInstallBanner';
@@ -14,5 +14,5 @@ redirectToLoginPage();
 wrapAndRenderLegacyCode(<AppInstallBanner />, 'footer-install-app');
 wrapAndRenderLegacyCode(<NotificationMenu />, 'notificationPanel');
 wrapAndRenderLegacyCode(<AppMenu />, 'appMenu');
-wrapAndRenderLegacyCode(<UserMenu />, 'userMenu');
+wrapAndRenderLegacyCode(<UserMenuAuthenticated />, 'userMenu');
 wrapAndRenderLegacyCode(<UserMenuUnauthenticated />, 'userMenuUnauthenticated');
