+++ frontend/src/shared/components/PageTemplate/AppMenu/components/AppMenuPanel.tsx
@@ -1,17 +1,8 @@
-import { NavLink } from 'react-router-dom';
-
-import {
-  Drawer,
-  Icon,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemIcon,
-  ListItemText,
-  Toolbar,
-} from '@mui/material';
+import { Drawer, List, Toolbar } from '@mui/material';
 
 import { useTranslation } from '#shared/i18n/useTranslation';
+
+import { AppMenuItem } from './AppMenuItem';
 
 interface AppMenuPanelProps {
   menuOpen: boolean;
@@ -82,43 +73,5 @@ export function AppMenuPanel({ menuOpen, onClose }: AppMenuPanelProps) {
         />
       </List>
     </Drawer>
-  );
-}
-
-interface AppMenuItemProps {
-  label: string;
-  path: string;
-  iconPath?: string;
-  isOnBackend?: boolean;
-  closeMenu: () => void;
-}
-
-function AppMenuItem({
-  label,
-  path,
-  iconPath = '/static/img/icons/cropped/link.svg',
-  isOnBackend,
-  closeMenu,
-}: AppMenuItemProps) {
-  return (
-    <ListItem disablePadding>
-      <ListItemButton
-        component={NavLink}
-        to={path}
-        reloadDocument={isOnBackend}
-        onClick={closeMenu}
-        sx={{
-          '&.active': {
-            color: 'primary.main',
-            '& span': { fontWeight: 500 },
-          },
-        }}
-      >
-        <ListItemIcon>
-          <Icon component="img" src={iconPath} alt="" />
-        </ListItemIcon>
-        <ListItemText>{label}</ListItemText>
-      </ListItemButton>
-    </ListItem>
   );
 }
