+++ frontend/src/shared/components/PageTemplate/NavBar/components/BreadcrumbsNav.tsx
@@ -1,10 +1,12 @@
-import { Link, useLocation, useMatches } from 'react-router-dom';
+import { useMatches } from 'react-router-dom';
 
 import { NavigateNext as NavigateNextIcon } from '@mui/icons-material';
-import { Breadcrumbs, Button, Icon, Typography } from '@mui/material';
+import { Breadcrumbs } from '@mui/material';
 
 import { useBreakpoint } from '#shared/hooks/useBreakpoint';
 import { useTranslation } from '#shared/i18n/useTranslation';
+
+import { BreadcrumbsItem } from './BreadcrumbsItem';
 
 export function BreadcrumbsNav() {
   const matches = useMatches();
@@ -52,62 +54,5 @@ export function BreadcrumbsNav() {
         />
       ))}
     </Breadcrumbs>
-  );
-}
-
-interface BreadcrumbsItemProps {
-  path: string;
-  label: string;
-  icon?: string;
-  fontWeight?: 'bold' | 'normal';
-  minimized?: boolean;
-}
-function BreadcrumbsItem({
-  path,
-  label,
-  icon,
-  fontWeight = 'normal',
-  minimized = false,
-}: BreadcrumbsItemProps) {
-  const { pathname } = useLocation();
-  const isCurrentPage = path.replace(/\/$/, '') === pathname.replace(/\/$/, '');
-
-  const scrollToTop = (e) => {
-    e.preventDefault();
-    window.scroll({ behavior: 'smooth', top: 0 });
-  };
-
-  return (
-    <Button
-      component={Link}
-      to={path}
-      onClick={isCurrentPage ? scrollToTop : undefined}
-      color="inherit"
-      sx={{
-        columnGap: 1,
-        textTransform: 'none',
-        minWidth: 'unset',
-        borderRadius: '100px',
-      }}
-    >
-      {icon && (
-        <Icon
-          component="img"
-          src={icon}
-          alt=""
-          sx={{ width: 32, height: 32 }}
-        />
-      )}
-      {!minimized && (
-        <Typography
-          variant="h6"
-          component="span"
-          fontWeight={fontWeight}
-          noWrap
-        >
-          {label}
-        </Typography>
-      )}
-    </Button>
   );
 }
